requirements.txt
aiodns==3.1.1
aiogram==2.25.1
aiohttp==3.8.3
aioredis==2.0.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.6.2
async-rediscache==0.2.0
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
blinker==1.7.0
branca==0.7.0
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
class-registry==2.1.2
click==8.1.7
colorama==0.4.6
discord==2.3.2
discord.py==2.3.2
distlib==0.3.6
filelock==3.9.0
filters==1.3.2
flake8==6.1.0
Flask==3.0.0
folium==0.15.1
frozenlist==1.3.3
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.25.1
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
magic-filter==1.0.9
MarkupSafe==2.1.3
mccabe==0.7.0
multidict==6.0.4
numpy==1.26.2
pipenv==2023.2.4
platformdirs==3.0.0
pycares==4.4.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydis_core==10.5.1
pyflakes==3.1.0
pyTelegramBotAPI==4.14.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-telegram-bot==20.6
pytz==2022.7.1
redis==5.0.1
regex==2023.10.3
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
statsd==4.0.1
telegram==0.0.1
typing_extensions==4.9.0
urllib3==2.0.7
virtualenv==20.19.0
virtualenv-clone==0.5.7
Werkzeug==3.0.1
xyzservices==2023.10.1
yarl==1.8.2